Add unit tests and some validations
